feat: add logout forgetPassword, setNewPassword endpoints,
